+++ .circleci/get-k6.sh
@@ -1,0 +1,21 @@
+#!/bin/bash
+#
+# Copyright 2018-2023 contributors to the Marquez project
+# SPDX-License-Identifier: Apache-2.0
+#
+# Usage: $ ./get-k6.sh
+
+set -e
+
+# Delete existing key (if present)
+sudo apt-key del k6
+
+# Add k6 key and update the repository
+sudo gpg -k && \
+  sudo gpg --no-default-keyring --keyring /usr/share/keyrings/k6-archive-keyring.gpg --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys C5AD17C747E3415A3642D57D77C6C491D6AC1D69
+echo "deb [signed-by=/usr/share/keyrings/k6-archive-keyring.gpg] https://dl.k6.io/deb stable main" | sudo tee /etc/apt/sources.list.d/k6.list
+
+# Install k6, then verify
+sudo snap install k6 && k6 version
+
+echo "DONE!"