+++ web/src/components/lineage/types.ts
@@ -1,3 +1,4 @@
+// Copyright 2018-2023 contributors to the Marquez project
 // SPDX-License-Identifier: Apache-2.0
 
 import { Field, Run, Tag } from '../../types/api'
@@ -6,6 +7,8 @@ import { Nullable } from '../../types/util/Nullable'
 export type JobOrDataset = 'JOB' | 'DATASET'
 export type BatchOrStream = 'BATCH' | 'STREAM' | 'SERVICE'
 export type DbTableOrStream = 'DB_TABLE' | 'STREAM'
+export type ariaJobLabel = string
+export type ariaDatasetLabel = string
 
 export interface LineageDataset {
   id: { namespace: string; name: string }