+++ web/src/store/sagas/index.ts
@@ -1,7 +1,11 @@
+// Copyright 2018-2023 contributors to the Marquez project
 // SPDX-License-Identifier: Apache-2.0
 
 import * as Effects from 'redux-saga/effects'
 import {
+  DELETE_DATASET,
+  DELETE_JOB,
+  FETCH_DATASET,
   FETCH_DATASETS,
   FETCH_DATASET_VERSIONS,
   FETCH_EVENTS,
@@ -17,6 +21,9 @@ const call: any = Effects.call
 
 import {
   applicationError,
+  deleteDatasetSuccess,
+  deleteJobSuccess,
+  fetchDatasetSuccess,
   fetchDatasetVersionsSuccess,
   fetchDatasetsSuccess,
   fetchEventsSuccess,
@@ -27,6 +34,9 @@ import {
   fetchSearchSuccess
 } from '../actionCreators'
 import {
+  deleteDataset,
+  deleteJob,
+  getDataset,
   getDatasetVersions,
   getDatasets,
   getEvents,
@@ -95,6 +105,18 @@ export function* fetchJobsSaga() {
   }
 }
 
+export function* deleteJobSaga() {
+  while (true) {
+    try {
+      const { payload } = yield take(DELETE_JOB)
+      const job = yield call(deleteJob, payload.jobName, payload.namespace)
+      yield put(deleteJobSuccess(job.name))
+    } catch (e) {
+      yield put(applicationError('Something went wrong while removing job'))
+    }
+  }
+}
+
 export function* fetchDatasetsSaga() {
   while (true) {
     try {
@@ -119,6 +141,30 @@ export function* fetchEventsSaga() {
   }
 }
 
+export function* fetchDatasetSaga() {
+  while (true) {
+    try {
+      const { payload } = yield take(FETCH_DATASET)
+      const datasets = yield call(getDataset, payload.namespace, payload.name)
+      yield put(fetchDatasetSuccess(datasets))
+    } catch (e) {
+      yield put(applicationError('Something went wrong while fetching dataset'))
+    }
+  }
+}
+
+export function* deleteDatasetSaga() {
+  while (true) {
+    try {
+      const { payload } = yield take(DELETE_DATASET)
+      const dataset = yield call(deleteDataset, payload.datasetName, payload.namespace)
+      yield put(deleteDatasetSuccess(dataset.name))
+    } catch (e) {
+      yield put(applicationError('Something went wrong while removing job'))
+    }
+  }
+}
+
 export function* fetchDatasetVersionsSaga() {
   while (true) {
     try {
@@ -137,10 +183,13 @@ export default function* rootSaga(): Generator {
     fetchJobsSaga(),
     fetchRunsSaga(),
     fetchDatasetsSaga(),
+    fetchDatasetSaga(),
     fetchDatasetVersionsSaga(),
     fetchEventsSaga(),
     fetchLineage(),
-    fetchSearch()
+    fetchSearch(),
+    deleteJobSaga(),
+    deleteDatasetSaga()
   ]
 
   yield all([...sagasThatAreKickedOffImmediately, ...sagasThatWatchForAction])